simple_sound files added
git-svn-id: svn://localhost/steel.new/trunk@385 99e027b2-380c-0410-a737-a5784e1a7bde

+++ steel/trunk/sources/audio/sound.h
@@ -31,7 +31,7 @@ public:
 	float pitch;
 	bool isLoop;
 	float rolloffFactor;
-	float sourceRelative;
+	int sourceRelative;
 };
 
 #endif
